git-svn-id: http://yagf.googlecode.com/svn/trunk@243 bb55e26e-657b-11de-982f-55b50260c994

+++ src/mainform.cpp
@@ -350,7 +350,7 @@ void MainForm::closeEvent(QCloseEvent *event)
     QXtUnixSignalCatcher::catcher()->disconnectUnixSugnals();
 }
 
-void MainForm::rotateImage(int deg)
+void MainForm::rotateImage(qreal deg)
 {
     if (imageLoaded) {
         graphicsInput->clearBlocks();
