+++ src/qsnippet.cpp
@@ -42,7 +42,7 @@ bool QSnippet::addFile(const QString &name, const QImage * image)
     return true;
 }
 
-void QSnippet::setRotation(int rot)
+void QSnippet::setRotation(qreal rot)
 {
     rotation = rot;
 }
