+++ src/qsnippet.h
@@ -33,7 +33,7 @@ class QSnippet : public QListWidgetItem
 public:
     explicit QSnippet(QListWidget *parent = 0);
     bool addFile(const QString &name, const QImage * image = NULL);
-    void setRotation(int rot);
+    void setRotation(qreal rot);
     int getRotation();
     void setScale(double sc);
     double getScale();
@@ -48,7 +48,7 @@ signals:
 public slots:
 private:
     BolockList blockList;
-    int rotation;
+    qreal rotation;
     double scale;
     QString name;
     QRect crop1;
